airflow-requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-bash==5.2.1
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyspark==3.5.0
